refactor: 클래스 담당 역할별로 패키지 재구성
각 클래스의 역할에 따라 패키지 구성을 재배치함.
